% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{addIntervals}
\alias{addIntervals}
\title{Plot the mixture gamma/normal fit}
\usage{
addIntervals(
  x,
  variable,
  length = 1,
  start = NULL,
  end = NULL,
  unit = NULL,
  truncate = NULL
)
}
\arguments{
\item{variable}{The vvariable of interest, i.e. the variable that is fitted to the model (e.g. "distanceToBottom").}

\item{length}{The number of intervals.}

\item{start, end}{Optional start and end points of the interval span.}

\item{unit}{Set this to "days" if you need day intervals.}

\item{truncate}{An optinal two element vector of trucation limits for the time variable. Values below the first element are grouped into one interval, and values above the second element are grouped into one interval.}

\item{dateVariable}{The variable defining the intervals.}
}
\description{
Plot the mixture gamma/normal fit
}
