% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dcastParTable}
\alias{dcastParTable}
\title{dcast a parTable}
\usage{
dcastParTable(
  parTable,
  dateVariable,
  timeVariable,
  variables = c("shape", "rate", "mu", "sigma", "p", "MeanMaxLogLikelihood")
)
}
\arguments{
\item{parTable}{The parTable returned from estimateDepthDistribution().}

\item{dateVariable}{The variable defining the date, defaulted to "Date".}

\item{timeVariable}{The variable defining the (diurnal) time, defaulted to "Hour".}

\item{variables}{A vector of parameters to display in the plots.}
}
\description{
dcast a parTable
}
