% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{estimateDepthDistribution}
\alias{estimateDepthDistribution}
\title{Plot the mixture gamma/normal fit}
\usage{
estimateDepthDistribution(
  data,
  variable = "distanceToBottom",
  dateVariable = "Date",
  timeVariable = "Hour",
  dateIntervalLength = 10,
  timeIntervalLength = 2,
  timeTruncate = NULL,
  p0 = 0.5,
  lower = 0.01,
  upper = 0.99,
  minLength = 20
)
}
\arguments{
\item{data}{A table og sandeel depth data.}

\item{variable}{The vvariable of interest, i.e. the variable that is fitted to the model (e.g. "distanceToBottom").}

\item{dateVariable}{The variable defining the date, defaulted to "Date".}

\item{timeVariable}{The variable defining the (diurnal) time, defaulted to "Hour".}

\item{dateIntervalLength}{The length of the date intervals.}

\item{timeIntervalLength}{The length of the time intervals.}

\item{timeTruncate}{An optinal two element vector of trucation limits for the time variable. Values below the first element are grouped into one interval, and values above the second element are grouped into one interval.}

\item{p0}{Initial value for the probability that a school is feeding.}

\item{lower, upper}{The lower and upper limits for the optimization of p.}

\item{minLength}{The minimum number of schools required for fitting, defaulted to 20.}
}
\description{
Plot the mixture gamma/normal fit
}
