% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{readSandeelData}
\alias{readSandeelData}
\title{Read sandeel school depth data}
\usage{
readSandeelData(filePath)
}
\arguments{
\item{filePath}{The path to the sandeel school depth data file.}
}
\description{
Read sandeel school depth data
}
