% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandeelDepthDistribution-package.R
\docType{package}
\name{SandeelDepthDistribution}
\alias{SandeelDepthDistribution}
\alias{SandeelDepthDistribution-package}
\title{Fits mixture gamma/normal distribution to sandeel school vertical distribution data.}
\description{
This package fits a mixture probability distribution to sandeel school vertical distribution data, where the sandeel are split into two behaviors, a resting behavior close to the sea bed modeled by the gamma distribution, and a feeding behavior higher in the water column modeled by the normal distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnejohannesholmin/SandeelDepthDistribution}
  \item Report bugs at \url{https://github.com/arnejohannesholmin/SandeelDepthDistribution/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

Other contributors:
\itemize{
  \item Norwegian Institute of Marine Research [copyright holder, funder]
}

}
