% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotModelFit}
\alias{plotModelFit}
\title{Plot the mixture gamma/normal fit}
\usage{
plotModelFit(
  fit,
  variable,
  showFit = TRUE,
  showAllFits = FALSE,
  dateIntevalVariable = "DateInterval",
  timeIntevalVariable = "HourInterval",
  arrange = FALSE,
  adds = NULL,
  crop = FALSE,
  byrow = TRUE,
  valuesToPrint = c("N", "p", "shape", "rate", "mu", "sigma"),
  addMain = TRUE,
  xlim = NULL
)
}
\arguments{
\item{fit}{The output of \code{estimateDepthDistribution}.}

\item{variable}{The vvariable of interest, i.e. the variable that is fitted to the model (e.g. "distanceToBottom").}

\item{showAllFits}{Logical: If TRUE, plot all fits in the iterative method, with the final fit highlighted.}

\item{dateIntevalVariable}{The variable defining the date intervals, defaulted to "DateInterval",}

\item{timeIntevalVariable}{The variable defining the (diurnal) time intervals, defaulted to "HourInterval",}

\item{arrange}{Logical: If TRUE, arrange the plots using ggpubr::ggarrange.}

\item{adds}{Optional ggplot object to add to each plot.}

\item{crop}{Logical: If TRUE, crop the plots of the date and time intervals with no valid fits.}

\item{byrow}{Logical: Used when arrange = TRUE. If TRUE (the default) arrange the plots by  row first and then column.}

\item{valuesToPrint}{A vector of strings naming the estimated parameters to print in the plot. Possible values are "N", "p", "shape", "rate", "scale", "mu" and "sigma".}

\item{addMain}{Logical: If TRUE, add the automatically generated main title.}

\item{xlim}{A vector of min and max of the limit of the x axis. NULL (the default) implies individual automatic xlim from 0 to the maximum for each plot, and NA implies 0 to the maximum across all plots.}
}
\value{
A list of ggplots or an arranged ggplot if arrange = TRUE.
}
\description{
Plot the mixture gamma/normal fit
}
