% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotModelFit}
\alias{plotModelFit}
\title{Plot the mixture gamma/normal fit}
\usage{
plotModelFit(
  fit,
  variable,
  showFit = TRUE,
  showAllFits = FALSE,
  dateIntevalVariable = "DateInterval",
  timeIntevalVariable = "HourInterval",
  arrange = FALSE,
  adds = NULL,
  crop = FALSE
)
}
\arguments{
\item{fit}{The output of \code{estimateDepthDistribution}.}

\item{variable}{The vvariable of interest, i.e. the variable that is fitted to the model (e.g. "distanceToBottom").}

\item{showAllFits}{Logical: If TRUE, plot all fits in the iterative method, with the final fit highlighted.}

\item{dateIntevalVariable}{The variable defining the date intervals, defaulted to "DateInterval",}

\item{timeIntevalVariable}{The variable defining the (diurnal) time intervals, defaulted to "HourInterval",}

\item{arrange}{Logical: If TRUE, arrange the plots using ggpubr::ggarrange.}

\item{adds}{Optional ggplot object to add to each plot.}

\item{crop}{Logical: If TRUE, crop the plots of the date and time intervals with no valid fits.}
}
\value{
A list of ggplots or an arranged ggplot if arrange = TRUE.
}
\description{
Plot the mixture gamma/normal fit
}
